moves endpoint into own file for structure consistency
